projects/api-automation/requirements.txt
psycopg2-binary
pytest==8.2.1
allure-pytest==2.13.5